finalproject (2)/finalproject/requirements.txt
Django
google-generativeai
python-dotenv
